get abstraction working for plotly
